+++ plugged/render-markdown.nvim/lua/render-markdown/render/heading.lua
@@ -1,4 +1,5 @@
 local Base = require('render-markdown.render.base')
+local Iter = require('render-markdown.core.iter')
 local colors = require('render-markdown.colors')
 local list = require('render-markdown.core.list')
 local str = require('render-markdown.core.str')
@@ -10,8 +11,17 @@ local str = require('render-markdown.core.str')
 ---@field sign? string
 ---@field foreground string
 ---@field background string
----@field heading_width render.md.heading.Width
+---@field width render.md.heading.Width
+---@field left_margin number
+---@field left_pad number
+---@field right_pad number
+---@field min_width integer
 ---@field end_row integer
+
+---@class render.md.width.Heading
+---@field margin integer
+---@field padding integer
+---@field content integer
 
 ---@class render.md.render.Heading: render.md.Renderer
 ---@field private heading render.md.Heading
@@ -42,11 +52,6 @@ function Render:setup()
         atx, level = true, str.width(self.info.text)
     end
 
-    local heading_width = self.heading.width
-    if type(heading_width) == 'table' then
-        heading_width = list.clamp(heading_width, level)
-    end
-
     self.data = {
         atx = atx,
         level = level,
@@ -54,11 +59,28 @@ function Render:setup()
         sign = list.cycle(self.heading.signs, level),
         foreground = list.clamp(self.heading.foregrounds, level),
         background = list.clamp(self.heading.backgrounds, level),
-        heading_width = heading_width,
+        width = Render.resolve(self.heading.width, level),
+        left_margin = Render.resolve(self.heading.left_margin, level),
+        left_pad = Render.resolve(self.heading.left_pad, level),
+        right_pad = Render.resolve(self.heading.right_pad, level),
+        min_width = Render.resolve(self.heading.min_width, level),
         end_row = self.info.end_row + (atx and 1 or 0),
     }
 
     return true
+end
+
+---@private
+---@generic T
+---@param values `T`|T[]
+---@param level integer
+---@return T
+function Render.resolve(values, level)
+    if type(values) == 'table' then
+        return list.clamp(values, level)
+    else
+        return values
+    end
 end
 
 function Render:render()
@@ -68,49 +90,55 @@ function Render:render()
     end
     self:background(width)
     self:border(width)
-    self:left_pad()
+    self:left_pad(width)
     self:conceal_underline()
 end
 
 ---@private
 ---@return integer
 function Render:icon()
+    local icon, highlight = self.data.icon, { self.data.foreground, self.data.background }
+
     if not self.data.atx then
-        if self.data.icon == nil then
+        if icon == nil then
             return 0
         end
-        local added = self.marks:add(true, self.info.start_row, self.info.start_col, {
-            end_row = self.info.end_row,
-            end_col = self.info.end_col,
-            virt_text = { { self.data.icon, { self.data.foreground, self.data.background } } },
-            virt_text_pos = 'inline',
-        })
-        return added and str.width(self.data.icon) or 0
+        local added = true
+        for row = self.info.start_row, self.data.end_row - 1 do
+            added = added
+                and self.marks:add(true, row, self.info.start_col, {
+                    end_row = row,
+                    end_col = self.info.end_col,
+                    virt_text = { { row == self.info.start_row and icon or str.pad(str.width(icon)), highlight } },
+                    virt_text_pos = 'inline',
+                })
+        end
+        return added and str.width(icon) or 0
     end
 
     -- For atx headings available width is level + 1 - concealed, where level = number of
     -- `#` characters, one is added to account for the space after the last `#` but before
     -- the  heading title, and concealed text is subtracted since that space is not usable
     local width = self.data.level + 1 - self.context:concealed(self.info)
-    if self.data.icon == nil then
+    if icon == nil then
         return width
     end
 
-    local padding = width - str.width(self.data.icon)
+    local padding = width - str.width(icon)
     if self.heading.position == 'inline' or padding < 0 then
         local added = self.marks:add(true, self.info.start_row, self.info.start_col, {
             end_row = self.info.end_row,
             end_col = self.info.end_col,
-            virt_text = { { self.data.icon, { self.data.foreground, self.data.background } } },
+            virt_text = { { icon, highlight } },
             virt_text_pos = 'inline',
             conceal = '',
         })
-        return added and str.width(self.data.icon) or width
+        return added and str.width(icon) + 1 or width
     else
         self.marks:add(true, self.info.start_row, self.info.start_col, {
             end_row = self.info.end_row,
             end_col = self.info.end_col,
-            virt_text = { { str.pad(padding) .. self.data.icon, { self.data.foreground, self.data.background } } },
+            virt_text = { { str.pad(padding) .. icon, highlight } },
             virt_text_pos = 'overlay',
         })
         return width
@@ -119,40 +147,45 @@ end
 
 ---@private
 ---@param icon_width integer
----@return integer
+---@return render.md.width.Heading
 function Render:width(icon_width)
-    if self.data.heading_width == 'block' then
-        local width = nil
-        if self.data.atx then
-            width = icon_width + self.context:width(self.info:sibling('inline'))
-        else
-            -- Account for icon in first row
-            local widths = vim.tbl_map(str.width, self.info:lines())
-            widths[1] = widths[1] + icon_width
-            width = vim.fn.max(widths)
-        end
-        width = self.heading.left_pad + width + self.heading.right_pad
-        return math.max(width, self.heading.min_width)
+    local text_width = nil
+    if self.data.atx then
+        text_width = self.context:width(self.info:sibling('inline'))
     else
-        return self.context:get_width()
+        text_width = vim.fn.max(Iter.list.map(self.info:lines(), str.width))
     end
+    local width = icon_width + text_width
+    local left_padding = self.context:resolve_offset(self.data.left_pad, width)
+    local right_padding = self.context:resolve_offset(self.data.right_pad, width)
+    width = math.max(left_padding + width + right_padding, self.data.min_width)
+    ---@type render.md.width.Heading
+    return {
+        margin = self.context:resolve_offset(self.data.left_margin, width),
+        padding = left_padding,
+        content = self.data.width == 'block' and width or self.context:get_width(),
+    }
 end
 
 ---@private
----@param width integer
+---@param width render.md.width.Heading
 function Render:background(width)
-    local win_col = width + self:indent(self.data.level)
+    local win_col, padding = 0, {}
+    if self.data.width == 'block' then
+        win_col = width.margin + width.content + self:indent(self.data.level)
+        table.insert(padding, { str.pad(vim.o.columns * 2), self.config.padding.highlight })
+    end
     for row = self.info.start_row, self.data.end_row - 1 do
         self.marks:add(true, row, 0, {
             end_row = row + 1,
             hl_group = self.data.background,
             hl_eol = true,
         })
-        if self.data.heading_width == 'block' then
+        if win_col > 0 and #padding > 0 then
             -- Overwrite anything beyond width with padding highlight
             self.marks:add(true, row, 0, {
                 priority = 0,
-                virt_text = { { str.pad(vim.o.columns * 2), self.config.padding.highlight } },
+                virt_text = padding,
                 virt_text_win_col = win_col,
             })
         end
@@ -160,22 +193,30 @@ function Render:background(width)
 end
 
 ---@private
----@param width integer
+---@param width render.md.width.Heading
 function Render:border(width)
     -- Only atx headings support borders
     if not self.heading.border or not self.data.atx then
         return
     end
 
-    local background = colors.inverse_bg(self.data.background)
+    local foreground, background = self.data.foreground, colors.inverse_bg(self.data.background)
     local prefix = self.heading.border_prefix and self.data.level or 0
+    local virtual = self.heading.border_virtual
 
-    local line_above = {
-        { self.heading.above:rep(self.heading.left_pad), background },
-        { self.heading.above:rep(prefix), self.data.foreground },
-        { self.heading.above:rep(width - self.heading.left_pad - prefix), background },
-    }
-    if str.width(self.info:line('above', 1)) == 0 and self.info.start_row - 1 ~= self.context.last_heading then
+    ---@param icon string
+    ---@return { [1]: string, [2]: string }[]
+    local function line(icon)
+        return {
+            { str.pad(width.margin), self.config.padding.highlight },
+            { icon:rep(width.padding), background },
+            { icon:rep(prefix), foreground },
+            { icon:rep(width.content - width.padding - prefix), background },
+        }
+    end
+
+    local line_above = line(self.heading.above)
+    if not virtual and self:empty_line('above') and self.info.start_row - 1 ~= self.context.last_heading then
         self.marks:add(true, self.info.start_row - 1, 0, {
             virt_text = line_above,
             virt_text_pos = 'overlay',
@@ -187,12 +228,8 @@ function Render:border(width)
         })
     end
 
-    local line_below = {
-        { self.heading.below:rep(self.heading.left_pad), background },
-        { self.heading.below:rep(prefix), self.data.foreground },
-        { self.heading.below:rep(width - self.heading.left_pad - prefix), background },
-    }
-    if str.width(self.info:line('below', 1)) == 0 then
+    local line_below = line(self.heading.below)
+    if not virtual and self:empty_line('below') then
         self.marks:add(true, self.info.end_row + 1, 0, {
             virt_text = line_below,
             virt_text_pos = 'overlay',
@@ -206,16 +243,30 @@ function Render:border(width)
 end
 
 ---@private
-function Render:left_pad()
-    if self.heading.left_pad <= 0 then
-        return
+---@param position 'above'|'below'
+---@return boolean
+function Render:empty_line(position)
+    return str.width(self.info:line(position, 1)) == 0
+end
+
+---@private
+---@param width render.md.width.Heading
+function Render:left_pad(width)
+    local virt_text = {}
+    if width.margin > 0 then
+        table.insert(virt_text, { str.pad(width.margin), self.config.padding.highlight })
     end
-    for row = self.info.start_row, self.data.end_row - 1 do
-        self.marks:add(false, row, 0, {
-            priority = 0,
-            virt_text = { { str.pad(self.heading.left_pad), self.data.background } },
-            virt_text_pos = 'inline',
-        })
+    if width.padding > 0 then
+        table.insert(virt_text, { str.pad(width.padding), self.data.background })
+    end
+    if #virt_text > 0 then
+        for row = self.info.start_row, self.data.end_row - 1 do
+            self.marks:add(false, row, 0, {
+                priority = 0,
+                virt_text = virt_text,
+                virt_text_pos = 'inline',
+            })
+        end
     end
 end